+++ CHANGES.md
@@ -8,7 +8,8 @@ CHANGES
   to `WebSocketListener`.
 - Added `DECOMPRESSION_ERROR` to `WebSocketError`.
 - Fixed a deadlock issue reported in #21.
-- Changed the frame queue in `WritingThread` from `List` to `Queue` for #21.
+- Changed the frame queue in `WritingThread` from `List` to `Deque` for #21.
+- Changed `WritingThread` to process PING & PONG frames immediately.
 
 
 1.15 (2015-12-25)